Fixed bug
If a provider was removed the provider count would be decremented one and this would cause the next provider added to have aduplicate ID unless the provider removed was at the end of the list.

If the ID's are compared and ever match the provider being added will have their ID incremented once and checked again so it does not match ever.